+++ lib/aixm.rb
@@ -48,6 +48,7 @@ require_relative 'aixm/component/runway'
 require_relative 'aixm/component/fato'
 require_relative 'aixm/component/helipad'
 require_relative 'aixm/component/surface'
+require_relative 'aixm/component/vasis'
 require_relative 'aixm/component/lighting'
 require_relative 'aixm/component/approach_lighting'