+++ spec/lib/aixm/document_spec.rb
@@ -264,6 +264,13 @@ describe AIXM::Document do
             <valMagBrg>164</valMagBrg>
             <valElevTdz>145</valElevTdz>
             <uomElevTdz>FT</uomElevTdz>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <codeVfrPattern>E</codeVfrPattern>
             <txtRmk>forth remarks</txtRmk>
           </Rdn>
@@ -339,6 +346,13 @@ describe AIXM::Document do
             <valMagBrg>344</valMagBrg>
             <valElevTdz>147</valElevTdz>
             <uomElevTdz>FT</uomElevTdz>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <codeVfrPattern>L</codeVfrPattern>
             <txtRmk>back remarks</txtRmk>
           </Rdn>
@@ -440,6 +454,13 @@ describe AIXM::Document do
             </FdnUid>
             <valTrueBrg>355</valTrueBrg>
             <valMagBrg>354</valMagBrg>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <txtRmk>Avoid flight over residental area</txtRmk>
           </Fdn>
           <Fls>
@@ -1143,6 +1164,13 @@ describe AIXM::Document do
             <valMagBrg>164</valMagBrg>
             <valElevTdz>145</valElevTdz>
             <uomElevTdz>FT</uomElevTdz>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <codeVfrPattern>E</codeVfrPattern>
             <txtRmk>forth remarks</txtRmk>
           </Rdn>
@@ -1218,6 +1246,13 @@ describe AIXM::Document do
             <valMagBrg>344</valMagBrg>
             <valElevTdz>147</valElevTdz>
             <uomElevTdz>FT</uomElevTdz>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <codeVfrPattern>L</codeVfrPattern>
             <txtRmk>back remarks</txtRmk>
           </Rdn>
@@ -1319,6 +1354,13 @@ describe AIXM::Document do
             </FdnUid>
             <valTrueBrg>355</valTrueBrg>
             <valMagBrg>354</valMagBrg>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <txtRmk>Avoid flight over residental area</txtRmk>
           </Fdn>
           <Fls>
@@ -2100,6 +2142,13 @@ describe AIXM::Document do
             <valMagBrg>164</valMagBrg>
             <valElevTdz>145</valElevTdz>
             <uomElevTdz>FT</uomElevTdz>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <codeVfrPattern>E</codeVfrPattern>
             <txtRmk>forth remarks</txtRmk>
           </Rdn>
@@ -2175,6 +2224,13 @@ describe AIXM::Document do
             <valMagBrg>344</valMagBrg>
             <valElevTdz>147</valElevTdz>
             <uomElevTdz>FT</uomElevTdz>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <codeVfrPattern>L</codeVfrPattern>
             <txtRmk>back remarks</txtRmk>
           </Rdn>
@@ -2276,6 +2332,13 @@ describe AIXM::Document do
             </FdnUid>
             <valTrueBrg>355</valTrueBrg>
             <valMagBrg>354</valMagBrg>
+            <codeTypeVasis>PAPI</codeTypeVasis>
+            <codePsnVasis>BOTH</codePsnVasis>
+            <noBoxVasis>2</noBoxVasis>
+            <codePortableVasis>N</codePortableVasis>
+            <valSlopeAngleGpVasis>6</valSlopeAngleGpVasis>
+            <valMeht>100</valMeht>
+            <uomMeht>FT</uomMeht>
             <txtRmk>Avoid flight over residental area</txtRmk>
           </Fdn>
           <Fls>