+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * `AIXM::R` (rectangle)
 * `Runway#marking`
 * `ApproachLighting` on `Runway::Direction` and `FATO::Direction`
+* `VASIS` on `Runway::Direction` and `FATO::Direction`
 
 #### Changes
 * Nested memoization of the same method is now allowed and won't reset the